Add keytab management functions (#10)
Adds the basic set of functions that are used to manage a keytab. These
functions allow the caller to enumerate a keytab as well as add/remote
any entries in a keytab. There are also some limited calls to inspect
the values of each entry.

Finally there are some APIs for creating a new keyblock that is required
when adding to a keytab. These APIs are fairly rudimentary. More
advanced APIs need to be added at a later date to make this more inline
with what the `ktutil` command can do.