Migrate to using PEP 517
Migrates the project to be PEP 517 compliant by specifying the build
metadata and other configuration in the pyproject.toml file. The only
remaining code in the setup.py is the native extensions which need to
run when installing from an sdist.

Part of this change is also removing the cythonised files from the sdist
making Cython a build requirement. As this is specified in the build
metadata this should automatically be installed when pip is installing
from an sdist.