Fix handling of krb5_data structs.
The caller (at least, MIT Kerberos) expects us to return our data in
an existing buffer referred to by the krb5_data struct we get, rather
than *update* that struct with a new buffer of our own.

I ran into this when using pykrb5 with PKINIT, and needed to supply a
passphrase for the client's private key via a Krb5Prompt object. It
mysteriously failed to decrypt the key even though the passphrase
returned by the prompter was right. Eventually I tracked it down to:

https://github.com/krb5/krb5/blob/30429ade54bfe66f9145a30487e43b19bde76701/src/plugins/preauth/pkinit/pkinit_crypto_openssl.c#L1196

This:

  rdat.data = buf;
  rdat.length = size

is where it sets the krb5_data rdat it will pass to us to refer to its
own buffer, and if you follow the code you see that it expects the
returned data to be in buf. We were instead returning an updated
rdat.data, which it never looks at.