Move back to using setup.cfg
Unfortunately due to some incompatibilities with pip and newer
setuptools versions that introduced support for project definitions in
the pyproject.toml this commit goes back to using the older setup.cfg
definition. This ensures the project can be compiled from the sdist in
more scenarios. At some point in the future this should be revisited
when hosts come with newer pip versions.